move gaspari_cohn and calc_geodist_exact to core
